stock_analysis/pyproject.toml
[tool.poetry]
name = "stock-analysis-crew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
crewai = "^0.11.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
tools = "^0.1.9"
wikipedia = "1.4.0"
yfinance = "0.2.35"
sec-api = ">=1.0.17"
tiktoken = "0.5.2"
faiss-cpu = ">=1.7.4"
python-dotenv = "1.0.0"
langchain-community = "0.0.10"
langchain-core = "0.1.8"
openai = "1.7.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"